Exploded WallabagRestController into many controllers Fix #2503
